Rename package for tests
