+++ src/main/java/ca/mcmaster/se2aa4/island/team037/actions/Action.java
@@ -1,9 +1,9 @@
-package ca.mcmaster.se2aa4.island.teamXXX.actions;
+package ca.mcmaster.se2aa4.island.team037.actions;
 
 import org.json.JSONObject;
 
-import ca.mcmaster.se2aa4.island.teamXXX.drone.Drone;
-import ca.mcmaster.se2aa4.island.teamXXX.result.ActionResult;
+import ca.mcmaster.se2aa4.island.team037.drone.Drone;
+import ca.mcmaster.se2aa4.island.team037.result.ActionResult;
 
 public interface Action {
     public ActionType type();
