+++ src/main/java/ca/mcmaster/se2aa4/island/team037/algorithm/State.java
@@ -1,8 +1,8 @@
-package ca.mcmaster.se2aa4.island.teamXXX.algorithm;
+package ca.mcmaster.se2aa4.island.team037.algorithm;
 
-import ca.mcmaster.se2aa4.island.teamXXX.actions.Action;
-import ca.mcmaster.se2aa4.island.teamXXX.drone.Drone;
-import ca.mcmaster.se2aa4.island.teamXXX.result.ActionResult;
+import ca.mcmaster.se2aa4.island.team037.actions.Action;
+import ca.mcmaster.se2aa4.island.team037.drone.Drone;
+import ca.mcmaster.se2aa4.island.team037.result.ActionResult;
 
 public abstract class State {
     private Drone drone;
