+++ src/main/java/ca/mcmaster/se2aa4/island/team037/drone/POIType.java
@@ -1,4 +1,4 @@
-package ca.mcmaster.se2aa4.island.teamXXX.drone;
+package ca.mcmaster.se2aa4.island.team037.drone;
 
 /**
  * Represents the type of a point of interest.
