+++ src/test/java/ca/mcmaster/se2aa4/island/team037/DimensionsTest.java
@@ -1,7 +1,9 @@
-package ca.mcmaster.se2aa4.island.teamXXX;
+package ca.mcmaster.se2aa4.island.team037;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import ca.mcmaster.se2aa4.island.team037.Explorer;
 
 public class DimensionsTest {
     private ExplorerFactory explorerFactory;
