+++ src/test/java/ca/mcmaster/se2aa4/island/team037/actions/FlyActionTest.java
@@ -1,14 +1,21 @@
-package ca.mcmaster.se2aa4.island.teamXXX.actions;
+package ca.mcmaster.se2aa4.island.team037.actions;
 
 import org.json.JSONObject;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
 
-import ca.mcmaster.se2aa4.island.teamXXX.result.ActionResult;
-import ca.mcmaster.se2aa4.island.teamXXX.drone.Drone;
-import ca.mcmaster.se2aa4.island.teamXXX.drone.Position;
-import ca.mcmaster.se2aa4.island.teamXXX.drone.Direction;
+import ca.mcmaster.se2aa4.island.team037.actions.ActionManager;
+import ca.mcmaster.se2aa4.island.team037.actions.ActionType;
+import ca.mcmaster.se2aa4.island.team037.actions.FlyAction;
+import ca.mcmaster.se2aa4.island.team037.drone.Direction;
+import ca.mcmaster.se2aa4.island.team037.actions.ActionManager;
+import ca.mcmaster.se2aa4.island.team037.actions.ActionType;
+import ca.mcmaster.se2aa4.island.team037.actions.FlyAction;
+import ca.mcmaster.se2aa4.island.team037.drone.Drone;
+import ca.mcmaster.se2aa4.island.team037.drone.Position;
+import ca.mcmaster.se2aa4.island.team037.result.ActionResult;
+
+import static org.junit.jupiter.api.Assertions.*;
 
 public class FlyActionTest {
     private Drone drone;
