+++ src/test/java/ca/mcmaster/se2aa4/island/team037/drone/DroneFactory.java
@@ -1,6 +1,8 @@
-package ca.mcmaster.se2aa4.island.teamXXX.drone;
+package ca.mcmaster.se2aa4.island.team037.drone;
 
-import ca.mcmaster.se2aa4.island.teamXXX.actions.ActionManager;
+import ca.mcmaster.se2aa4.island.team037.actions.ActionManager;
+import ca.mcmaster.se2aa4.island.team037.drone.Direction;
+import ca.mcmaster.se2aa4.island.team037.drone.Drone;
 
 public class DroneFactory {
     public Drone createDrone(Direction initialDirection, int batteryLevel, ActionManager actionManager) {
