+++ src/test/java/ca/mcmaster/se2aa4/island/team037/result/ScanActionResultFactory.java
@@ -1,7 +1,11 @@
-package ca.mcmaster.se2aa4.island.teamXXX.result;
+package ca.mcmaster.se2aa4.island.team037.result;
 
 import java.util.Arrays;
 import java.util.List;
+
+import ca.mcmaster.se2aa4.island.team037.result.ScanActionResult;
+
+import ca.mcmaster.se2aa4.island.team037.result.ScanActionResult;
 
 public class ScanActionResultFactory {
     public List<String> createCreeks() {
